kb control to gamecontrol.c
