+++ src/gamecontrol.h
@@ -1,0 +1,10 @@
+#ifndef GAMECONTROL_H_
+#define GAMECONTROL_H_
+/*
+*/
+    #include <math.h>
+    #define _USE_MATH_DEFINES
+    #include "draw.h"
+    #include "gamecontrol.c"
+
+#endif //GAMECONTROL_H_
