+++ src/main.c
@@ -27,7 +27,9 @@ int main(int argc, char const *argv[]) {
     initOS();
     
     while(loop) {
+
         while(getKeysInBuffer()) player_loc = kb_control(player_loc); // inBuffer > 0
+        
         if (!isMenu){
             ray_cast(player_loc, distance);
 
@@ -45,64 +47,18 @@ int main(int argc, char const *argv[]) {
             fprintB(fb, "time: %10.4f ms; fps: %10.0f; frame: %10.0lu; \n", (float) tTaken / 1000, (float) 1.0/tTaken*1000000, frame); 
             tLast = tNow;
             frame++;
-        }else draw_menu(fb);
+        }
+
+        else draw_menu(fb);
 
         displayB(fb, &codes[0], UNILEN); // write buffer to cli
     }
-    return 0;
+    return 0; 
 }
 
 
 
-struct position kb_control(struct position player_loc)
-{
-    struct position new;
-    new = player_loc;
-    int key = getKey(); // get character
 
-    if (!isMenu){
-        float pRad = player_loc.pA*M_PI/180; // degree to radians
-        switch (key) {
-            // rotate or walk according to key press, exit if '.'
-            case 'w':
-                new.pX += sin(pRad)*stepXY;
-                new.pY += cos(pRad)*stepXY;
-                break;
-            case 'a':
-                player_loc.pA+=stepA;
-                break;
-            case 's':
-                new.pX -= sin(pRad)*stepXY;
-                new.pY -= cos(pRad)*stepXY;
-                break;
-            case 'd':
-                player_loc.pA-=stepA;
-                break;
-            case  '.':
-                isMenu = 1;
-                break;
-        }
-    }
-    else {
-        switch (key) {
-            case '.':
-                loop = 0;
-                break;
-            case ' ':
-                isMenu = 0;
-                break;
-        }
-    }
-    // check if theoretical new position is not in wall -> write it to current position
-    // separate check for walk along wall
-    if(lvl.world[(int)(player_loc.pY)][(int)new.pX] == 0) {
-        player_loc.pX = new.pX;
-    }
-    if(lvl.world[(int)(new.pY)][(int)player_loc.pX] == 0) {
-        player_loc.pY = new.pY;
-    }
-    return player_loc;
-}
 void ray_cast(struct position player_loc, unsigned short *distance){
     // ray casting:
     // draw lines from player_loc in every direction in view
