+++ src/quest.h
@@ -4,6 +4,7 @@
 #include <sys/time.h>
 #include "crossplatform.h"
 #include <math.h>
+#include "gamecontrol.h"
 #include "framebuffer.h"
 #include "draw.h"
